2008-03-09  Tatsuhiro Tsujikawa  <tujikawa at rednoah dot com>
	type cleanup for allmost all remaining files.

+++ src/AbstractDiskWriter.h
@@ -45,34 +45,34 @@ class Logger;
 class AbstractDiskWriter : public DiskWriter {
 protected:
   std::string filename;
-  int32_t fd;
+  int fd;
   const Logger* logger;
 
-  void createFile(const std::string& filename, int32_t addFlags = 0);
+  void createFile(const std::string& filename, int addFlags = 0);
 
 private:
-  int32_t writeDataInternal(const unsigned char* data, int32_t len);
-  int32_t readDataInternal(unsigned char* data, int32_t len);
+  ssize_t writeDataInternal(const unsigned char* data, size_t len);
+  ssize_t readDataInternal(unsigned char* data, size_t len);
 
-  void seek(int64_t offset);
+  void seek(off_t offset);
 
 public:
   AbstractDiskWriter();
   virtual ~AbstractDiskWriter();
 
-  virtual void openFile(const std::string& filename, int64_t totalLength = 0);
+  virtual void openFile(const std::string& filename, uint64_t totalLength = 0);
 
   virtual void closeFile();
 
-  virtual void openExistingFile(const std::string& filename, int64_t totalLength = 0);
+  virtual void openExistingFile(const std::string& filename, uint64_t totalLength = 0);
 
-  virtual void writeData(const unsigned char* data, int32_t len, int64_t offset);
+  virtual void writeData(const unsigned char* data, size_t len, off_t offset);
 
-  virtual int32_t readData(unsigned char* data, int32_t len, int64_t offset);
+  virtual ssize_t readData(unsigned char* data, size_t len, off_t offset);
 
-  virtual void truncate(int64_t length);
+  virtual void truncate(uint64_t length);
 
-  virtual int64_t size() const;
+  virtual uint64_t size() const;
   
   virtual void enableDirectIO();
