+++ src/AnnounceList.h
@@ -61,7 +61,7 @@ public:
   void reconfigure(const MetaEntry* announceListEntry);
   void reconfigure(const std::string& url);
 
-  int32_t countTier() const {
+  size_t countTier() const {
     return tiers.size();
   }
 
@@ -102,12 +102,12 @@ public:
   /**
    * Counts the number of tiers to which the "stopped" event can be sent.
    */
-  int32_t countStoppedAllowedTier() const;
+  size_t countStoppedAllowedTier() const;
 
   /**
    * Counts the number of tiers to which the "completed" event can be sent.
    */
-  int32_t countCompletedAllowedTier() const;
+  size_t countCompletedAllowedTier() const;
 
   /**
    * Moves current tier pointer to the tier to which the "stopped" event can
