+++ src/BtInteractive.h
@@ -60,7 +60,7 @@ public:
 
   virtual void sendPendingMessage() = 0;
 
-  virtual int32_t countPendingMessage() = 0;
+  virtual size_t countPendingMessage() = 0;
 
   virtual bool isSendingMessageInProgress() = 0;
