+++ src/BtKeepAliveMessage.h
@@ -47,7 +47,7 @@ class BtKeepAliveMessage : public SimpleBtMessage {
 private:
   unsigned char* msg;
 
-  static int32_t MESSAGE_LENGTH;
+  static size_t MESSAGE_LENGTH;
 public:
   BtKeepAliveMessage():msg(0) {}
 
@@ -55,15 +55,15 @@ public:
     delete [] msg;
   }
 
-  static const int8_t ID = 99;
+  static const uint8_t ID = 99;
 
-  virtual int8_t getId() { return ID; }
+  virtual uint8_t getId() { return ID; }
 
   virtual void doReceivedAction() {}
 
   virtual const unsigned char* getMessage();
 
-  virtual int32_t getMessageLength();
+  virtual size_t getMessageLength();
 
   virtual std::string toString() const {
     return "keep alive";
