+++ src/BtMessageDispatcher.h
@@ -56,7 +56,7 @@ public:
 
   virtual void sendMessages() = 0;
 
-  virtual void doCancelSendingPieceAction(int32_t index, int32_t begin, int32_t length) = 0;
+  virtual void doCancelSendingPieceAction(size_t index, uint32_t begin, size_t length) = 0;
 
   virtual void doCancelSendingPieceAction(const SharedHandle<Piece>& piece) = 0;
 
@@ -70,13 +70,13 @@ public:
 
   virtual bool isSendingInProgress() = 0;
 
-  virtual int32_t countMessageInQueue() = 0;
+  virtual size_t countMessageInQueue() = 0;
 
-  virtual int32_t countOutstandingRequest() = 0;
+  virtual size_t countOutstandingRequest() = 0;
 
-  virtual bool isOutstandingRequest(int32_t index, int32_t blockIndex) = 0;
+  virtual bool isOutstandingRequest(size_t index, size_t blockIndex) = 0;
 
-  virtual RequestSlot getOutstandingRequest(int32_t index, int32_t begin, int32_t length) = 0;
+  virtual RequestSlot getOutstandingRequest(size_t index, uint32_t begin, size_t length) = 0;
 
   virtual void removeOutstandingRequest(const RequestSlot& slot) = 0;
