+++ src/BtNotInterestedMessage.h
@@ -47,7 +47,7 @@ class BtNotInterestedMessage : public SimpleBtMessage {
 private:
   unsigned char* msg;
   
-  static int32_t MESSAGE_LENGTH;
+  static size_t MESSAGE_LENGTH;
 public:
   BtNotInterestedMessage():msg(0) {}
 
@@ -55,17 +55,17 @@ public:
     delete [] msg;
   }
 
-  static const int8_t ID = 3;
+  static const uint8_t ID = 3;
 
-  static BtNotInterestedMessageHandle create(const unsigned char* data, int32_t dataLength);
+  static BtNotInterestedMessageHandle create(const unsigned char* data, size_t dataLength);
 
-  virtual int8_t getId() { return ID; }
+  virtual uint8_t getId() { return ID; }
 
   virtual void doReceivedAction();
 
   virtual const unsigned char* getMessage();
 
-  virtual int32_t getMessageLength();
+  virtual size_t getMessageLength();
 
   virtual std::string toString() const;
